requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1.7
django-baton==2.6.2
django-bootstrap4==23.1
django-crispy-forms==2.0
django-extensions==3.2.1
django-paypal==2.0
django-phonenumber-field==7.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.5
flake8==6.0.0
gprof2dot==2022.7.29
graphviz==0.20.1
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
phonenumbers==8.13.8
Pillow==9.4.0
platformdirs==3.1.0
pycodestyle==2.10.0
pycparser==2.21
pydot==1.4.2
pyflakes==3.0.1
PyJWT==2.6.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
soupsieve==2.4.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
webargs==8.2.0
